pyproject.toml
[tool.poetry]
name = "spectral-to-splunk"
version = "0.1.0"
description = ""
authors = ["Jonas Rosenboom <jrosenboom@checkpoint.com>"]

[tool.poetry.scripts]
spectral_to_splunk = 'spectral_to_splunk.main:main'

[tool.poetry.dependencies]
python = "^3.7.7"
requests = "^2.28.1"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
